Add files via upload
Let's say we want to create a class that can be used to represent people at a university (e.g. faculty, staff, undergraduate students, and graduate students - at least 4 groups, can be different). For all types of people, we want to store their name and university ID. For specific types of people, we want to store additional information. For faculty, we want to store their title (i.e. Assistant Professor, Associate Professor, or Full Professor). For undergraduate students, we want to store their year of study (i.e. 1, 2, 3, or 4). For graduate students, we want to store their supervisor's university ID. Design a class hierarchy that can be used to represent the above collection of people.

Create a menu that:

1. Allows you to display all people and separately for each group (e.g. faculty, students, etc.)
2. Add a person
3. Edit the details of a specific person
4. Delete a specific person